delete dead code and add route for add_phrase
