+++ app.py
@@ -1,44 +1,43 @@
+import os
 import logging
 logging.basicConfig(level=logging.DEBUG)
 
-import os
+from flask import Flask, request, Response
 from slack import WebClient
 from slack.errors import SlackApiError
+from slackeventsapi import SlackEventAdapter
 
-BOT_TOKEN = os.environ.get('BOT_TOKEN')
-client = WebClient(BOT_TOKEN)
+#instantiate app
+app = Flask(__name__)
 
-def list_channels():
-  channels_call = client.conversations_list()
-  if channels_call.get('ok'):
-    return channels_call['channels']
-  return None
+events_adapter = SlackEventAdapter(
+    os.environ['SLACK_SIGNING_SECRET'], endpoint="/events", server=app)
+client = WebClient(os.environ['BOT_TOKEN'])
 
-def channel_info(channel_id):
-  channel_info = client.conversations_info(channel=channel_id)
-  if channel_info:
-    return channel_info['channel']
-  return None
+phrases = []
 
-def send_message(channel_id, message):
+def send_message(id, message):
   try:
     response = client.chat_postMessage(
-      channel=channel_id,
+      channel=id,
       text=message
     )
-    print(response)
   except SlackApiError as e:
     assert e.response["error"]
 
-if __name__ == '__main__':
-  channels = list_channels()
-  if channels:
-    print("Channels: ")
-    for channel in channels:
-      if channel['name'] == 'general':
-        print(channel['name'])
-        print(channel['id'])
-        send_message(channel['id'], "Hi, I am Bot")
-      print('-------')
-  else:
-    print("Unable to authenticate.")
+@events_adapter.on("message")
+def handle_message(event_data):
+  message = event_data["event"]
+  if "yolo" in message.get('text'):
+    user = message["user"]
+    message = "Hello %s!" % user
+    send_message(user, message)
+
+#route for adding a watch-phrase
+@app.route('/add_phrase', methods=['POST'])
+def add_phrase():
+  return Response('You added a phrase!')
+
+@app.route('/', methods=['GET'])
+def test():
+  return Response('It works!')
